toi uu toc do phan hoi

+++ backend_python/src/models.py
@@ -5,4 +5,4 @@ class NodeProgress(BaseModel):
     user_id: str
     node_id: str
     opened: bool
-    score: int | None = None
+    score: float | None = None
